Simplify HTTP outcall testing logic
