+++ Cargo.toml
@@ -15,9 +15,6 @@ opt-level = 'z'
 [profile.canister-release]
 inherits = "release"
 
-[features]
-mock = []
-
 [dependencies]
 candid = { workspace = true }
 ic-canister-log = { workspace = true }
@@ -42,7 +39,6 @@ hex = "0.4"
 futures = "0.3"
 
 [dev-dependencies]
-evm_rpc = { path = ".", features = ["mock"] }
 ic-ic00-types = { git = "https://github.com/dfinity/ic", rev = "release-2023-09-27_23-01" }
 ic-base-types = { git = "https://github.com/dfinity/ic", rev = "release-2023-09-27_23-01" }
 ic-config = { git = "https://github.com/dfinity/ic", rev = "release-2023-09-27_23-01" }
