+++ src/http.rs
@@ -1,4 +1,3 @@
-use candid::Decode;
 use cketh_common::eth_rpc::{HttpOutcallError, ProviderError, RpcError, ValidationError};
 use ic_canister_log::log;
 use ic_cdk::api::{
@@ -66,7 +65,7 @@ pub async fn do_http_request(
     inc_metric_entry!(host_requests, host.to_string());
     let mut request_headers = vec![HttpHeader {
         name: CONTENT_TYPE_HEADER.to_string(),
-        value: "application/json".to_string(),
+        value: CONTENT_TYPE_VALUE.to_string(),
     }];
     request_headers.extend(api.headers);
     let request = CanisterHttpRequestArgument {
@@ -118,204 +117,9 @@ pub async fn perform_http_request(
     request: CanisterHttpRequestArgument,
     #[allow(unused_variables)] cycles: u128,
 ) -> CallResult<HttpResponse> {
-    #[cfg(any(feature = "mock", test))]
-    {
-        if let Some(mock) = mock_http::MOCK_OUTCALL.with(|mock| mock.borrow_mut().take()) {
-            mock.assert_matches(&request);
-            let mut response = mock.response;
-            if let Some(transform) = request.transform {
-                let method = transform.function.0.method;
-                response = match method.as_str() {
-                    "__transform_json_rpc" => do_transform_http_request(
-                        Decode!(&transform.context, TransformArgs).unwrap(),
-                    ),
-                    _ => panic!("Unsupported transform: {}", method),
-                }
-            }
-            Ok(response)
-        } else {
-            panic!("No mock configured for HTTP request")
-        }
-    }
-    #[cfg(not(any(feature = "mock", test)))]
-    {
-        Ok(
-            ic_cdk::api::management_canister::http_request::http_request(request, cycles)
-                .await?
-                .0,
-        )
-    }
-}
-
-#[cfg(any(feature = "mock", test))]
-pub use mock_http::*;
-
-#[cfg(any(feature = "mock", test))]
-pub mod mock_http {
-    use std::cell::RefCell;
-
-    use ic_cdk::api::management_canister::http_request::{
-        CanisterHttpRequestArgument, HttpHeader, HttpMethod, HttpResponse,
-    };
-    thread_local! {
-        pub static MOCK_OUTCALL: RefCell<Option<MockOutcall>> = RefCell::new(None);
-    }
-
-    pub struct MockOutcallBody(pub Vec<u8>);
-
-    impl From<String> for MockOutcallBody {
-        fn from(string: String) -> Self {
-            string.as_bytes().to_vec().into()
-        }
-    }
-    impl<'a> From<&'a str> for MockOutcallBody {
-        fn from(string: &'a str) -> Self {
-            string.to_string().into()
-        }
-    }
-    impl From<Vec<u8>> for MockOutcallBody {
-        fn from(bytes: Vec<u8>) -> Self {
-            MockOutcallBody(bytes)
-        }
-    }
-
-    pub struct MockOutcallBuilder(MockOutcall);
-
-    impl MockOutcallBuilder {
-        pub fn new(status: u16, body: impl Into<MockOutcallBody>) -> Self {
-            Self(MockOutcall {
-                method: None,
-                url: None,
-                request_headers: None,
-                request_body: None,
-                response: HttpResponse {
-                    status: status.into(),
-                    headers: vec![],
-                    body: body.into().0,
-                },
-            })
-        }
-
-        pub fn method(mut self, method: HttpMethod) -> Self {
-            self.0.method = Some(method);
-            self
-        }
-
-        pub fn expect_url(mut self, url: impl ToString) -> Self {
-            self.0.url = Some(url.to_string());
-            self
-        }
-
-        pub fn expect_headers(mut self, headers: Vec<HttpHeader>) -> Self {
-            self.0.request_headers = Some(headers);
-            self
-        }
-
-        pub fn expect_body(mut self, body: impl Into<MockOutcallBody>) -> Self {
-            self.0.request_body = Some(body.into().0);
-            self
-        }
-
-        pub fn build(self) -> MockOutcall {
-            self.0
-        }
-    }
-
-    #[derive(Clone, Debug)]
-    pub struct MockOutcall {
-        pub method: Option<HttpMethod>,
-        pub url: Option<String>,
-        pub request_headers: Option<Vec<HttpHeader>>,
-        pub request_body: Option<Vec<u8>>,
-        pub response: HttpResponse,
-    }
-
-    impl MockOutcall {
-        pub fn mock_once(self) {
-            mock_http_request(self)
-        }
-
-        pub fn assert_matches(&self, request: &CanisterHttpRequestArgument) {
-            if let Some(ref url) = self.url {
-                assert_eq!(url, &request.url);
-            }
-            if let Some(ref headers) = self.request_headers {
-                assert_eq!(headers, &request.headers);
-            }
-            if let Some(ref body) = self.request_body {
-                assert_eq!(body, &request.body.as_deref().unwrap_or_default());
-            }
-        }
-    }
-
-    impl From<HttpResponse> for MockOutcall {
-        fn from(response: HttpResponse) -> Self {
-            Self {
-                method: None,
-                url: None,
-                request_headers: None,
-                request_body: None,
-                response,
-            }
-        }
-    }
-
-    pub fn assert_no_mock_http_request() {
-        assert!(
-            MOCK_OUTCALL.with(|mock| mock.borrow().is_none()),
-            "Previous mock HTTPS outcall was not used"
-        )
-    }
-
-    fn mock_http_request(mock: MockOutcall) {
-        assert_no_mock_http_request();
-        MOCK_OUTCALL.with(|current_mock| {
-            let mut current_mock = current_mock.borrow_mut();
-            *current_mock = Some(mock)
-        })
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use assert_matches::assert_matches;
-    use cketh_common::eth_rpc_client::providers::RpcApi;
-    use futures::executor::block_on;
-    use ic_cdk::api::management_canister::http_request::HttpHeader;
-
-    use crate::*;
-
-    #[test]
-    fn test_do_http_request() {
-        let principal = Principal::anonymous();
-        do_authorize(principal, Auth::Rpc);
-        do_authorize(principal, Auth::FreeRpc);
-
-        let payload = r#"{"id":1,"jsonrpc":"2.0","method":"eth_gasPrice","params":null}"#;
-        let expected_result = r#"{"id":1,"jsonrpc":"2.0","result":"0x00112233"}"#;
-        let url = "https://cloudflare-eth.com";
-        let headers = vec![HttpHeader {
-            name: CONTENT_TYPE_HEADER.to_string(),
-            value: "application/json".to_string(),
-        }];
-        MockOutcallBuilder::new(200, expected_result)
-            .expect_url(url)
-            .expect_body(payload)
-            .expect_headers(headers.clone())
-            .build()
-            .mock_once();
-
-        assert_matches!(
-            block_on(do_http_request(
-                principal,
-                ResolvedSource::Api(RpcApi {
-                    url: url.to_string(),
-                    headers
-                }),
-                payload,
-                1000
-            )),
-            Ok(_)
-        );
-    }
+    Ok(
+        ic_cdk::api::management_canister::http_request::http_request(request, cycles)
+            .await?
+            .0,
+    )
 }
